% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ssh_command.R
\name{create_ssh_command}
\alias{create_ssh_command}
\title{A function to build an ssh command}
\usage{
create_ssh_command(private_key_path = NA, target_port = 5431,
  local_port_for_forward = "10.200.10.1:5432",
  forward_target_host = "oliver-replica.criploulbgnu.us-west-2.rds.amazonaws.com",
  forward_target_port = 5432, reverse_tunnel_target_user = NA,
  reverse_tunnel_target = "52.90.57.218")
}
\arguments{
\item{private_key_path}{A locally accesible path to your previously registered private key.}

\item{target_port}{The target port number. Defaults to 5431.}

\item{local_port_for_forward}{The local port to be forwarded. Defaults to 5432.}

\item{forward_target_host}{The actual location of the oliver replica. Defaults to \code{oliver-replica.criploulbgnu.us-west-2.rds.amazonaws.com}.}

\item{forward_target_port}{The targeted port on the target host. Defaults to 5432.}

\item{reverse_tunnel_target_user}{A valid username on the reverse tunnel target.}

\item{reverse_tunnel_target}{The name of the reverse tunnel target. Defaults to \code{52.90.57.218}.}
}
\value{
None (invisible NULL).
}
\description{
Create an ssh command string to access the oliver replica database on AWS via reverse tunnel to a linux box administered by POC. 
This command string can then be interactively pasted into a terminal to set a connection to the oliver replica on your localhost.
}
\details{
The following dependencies are required in order to create a valid string:

\itemize{
  \item A public/private key pair with the public key shared to \email{mattbro@uw.edu} and residing on the reverse tunnel target.
  \item A valid username on the reverse tunnel target. Can be set by contacting \email{mattbro@uw.edu}.
}

Access to the oliver replica requires two-factor authentication. Authy (\url{https://www.authy.com/}) is used for this purpose. 
Authy account information should be forwarded to \email{mattbro@uw.edu} prior to trying to establish a connection to the replica with the command string or access will be denied.
}
\examples{
create_ssh_command(private_key_path = '/Users/mienkoja/.ssh/id_rsa'
                   ,reverse_tunnel_target_user = 'mienkoja')
                   
}
\seealso{
\code{\link{create_flat_files}}
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
