% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_var_attribute.R
\name{define_var_attribute}
\alias{define_var_attribute}
\title{Define a tibble attribute variable}
\usage{
define_var_attribute(data, population_member_id, value)
}
\description{
Define a tibble attribute variable
}
