% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_var_event.R
\name{define_var_event}
\alias{define_var_event}
\title{Define a tibble event variable}
\usage{
define_var_event(data, id, value_date, tz = "America/Los_Angeles")
}
\description{
Define a tibble event variable
}
