% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_varset.R
\name{define_varset}
\alias{define_varset}
\title{Define a tibble period varset}
\usage{
define_varset(var_defs = list(), population_member_id = "id_referral_visit",
  multi_join_type = "inner")
}
\arguments{
\item{var_defs}{A list of tibble variables (events, periods, or attributes).}

\item{population_member_id}{A unique identifier for each 'member' of the population of interest. Must be present in each tibble.}

\item{multi_join_type}{The name of the join you wish to make. Currently limited to \code{'inner'} or \code{'left'}. Defaults is \code{'inner'}.}
}
\description{
A function to join a list of tibbles, by a single \code{population_member_id}, using either the \code{inner_join} or \code{left_join} functions from \pkg{dplyr}.
}
