% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimateCellProps}
\alias{estimateCellProps}
\title{Calculates estimated cell-type proportions from sequencing-based data, using R package meffil to do the heavy lifting}
\usage{
estimateCellProps(bs, reference = NULL)
}
\arguments{
\item{bs}{A bsseq object. Removing low-coverage CpGs is not required}

\item{reference}{Reference used for cell-type estimations. Please reference meffil.list.cell.type.references() for the available references}
}
\description{
Calculates estimated cell-type proportions from sequencing-based data, using R package meffil to do the heavy lifting
}
\examples{
estimateCellProps(bs)
}
