% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimateTL}
\alias{estimateTL}
\title{Estimate telomere length from a bsseq object, using R package methylclock to do the heavy lifting}
\usage{
estimateTL(bs)
}
\arguments{
\item{bs}{A bsseq object. Removing low-coverage CpGs is not required}
}
\description{
Estimate telomere length from a bsseq object, using R package methylclock to do the heavy lifting
}
\examples{
estimateTL(bs)
}
