% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_for_data.R
\docType{data}
\name{Fatty_acids}
\alias{Fatty_acids}
\title{Sample dataset of fatty acids of olive trees}
\format{Dataset with 169 observation and 14 columns.

\describe{
\item{Sample_ID}{column with the name of every sample}
\item{Label}{Label of the geographic site of sampling}
\item{UTM_Est & UTM_Nord}{Geographic coordinates of the olive trees}
}}
\usage{
Fatty_acids or data(Fatty_acids) if the package is
  not loaded.
}
\description{
The dataset is the same that you can find in the CSV_data folder
  of the package. You can load this dataset directly in R with the
  \code{\link{data}} function to test the results obtained with the shiny app
  or to compute new and different analyzes.
}
\keyword{datasets}
