% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekers.R
\name{findICRs}
\alias{findICRs}
\title{Identify imprinted control regions (ICRs) from a bsseq object}
\usage{
findICRs(
  bs,
  minCov = 1,
  minSample = 0.5,
  numCG = 5,
  tissueType = NULL,
  lowerLimitSomatic = 0.35,
  upperLimitSomatic = 0.65,
  lowerLimitGamete = 0.1,
  upperLimitGamete = 0.8
)
}
\arguments{
\item{bs}{A bsseq object}

\item{minCov}{Minimum coverage of a CpG from a sample to be considered for further analysis. Default is 1}

\item{minSample}{Minimum portion of samples to meet minCov cutoff to be considered for further analysis. Default is 0.5 (50\%)}

\item{numCG}{Minimum number of adjacent CpGs used to be tested in the sliding window to identify ICRs. Default is 5}

\item{tissueType}{Type of tissue of samples. One of "somatic" or "gamete". Default is "somatic".}

\item{lowerLimitSomatic}{Lower limit of methylation percentage to use to determine hemimethylation, if tissue is somatic. Default is 0.35 (35\%). Note, this cutoff is recommended for somatic tissues.}

\item{upperLimitSomatic}{Lower limit of methylation percentage to use to determine hemimethylation, if tissue is somatic. Default is 0.65 (65\%). Note, this cutoff is recommended for somatic tissues.}

\item{lowerLimitGamete}{Lower limit of methylation percentage to use to determine hemimethylation, if tissue is gametic. Default is 0.35 (35\%). Note, this cutoff is recommended for somatic tissues.}

\item{upperLimitGamete}{Lower limit of methylation percentage to use to determine hemimethylation, if tissue is gametic. Default is 0.65 (65\%). Note, this cutoff is recommended for somatic tissues.}
}
\description{
Identify imprinted control regions (ICRs) from a bsseq object
}
\examples{
findICRs(bs, minCov = 1, minSample = 0.5, numCG = 5, lowerLimit = 0.35, upperLimit = 0.65)
}
