% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekers.R
\name{findLMRs}
\alias{findLMRs}
\title{Identify low methylated regions (LMRs) (and unmethylated regions (UMRs)) for each sample from a bsseq object}
\usage{
findLMRs(
  bs,
  minCov = 10,
  methCutoff = 0.5,
  numCG = 4,
  chrSelect = "chr22",
  umrs = FALSE,
  nCores = 1
)
}
\arguments{
\item{bs}{A bsseq object}

\item{minCov}{Minimum coverage by all samples across a CpG to be considered for further analysis. Default is 10}

\item{methCutoff}{Methylation value cutoff (from 0 to 1) used to look for LMRs/UMRs. Default is 0.5 (50\%)}

\item{numCG}{Minimum number of CpGs used to find LMRs/UMRs. Default is 4}

\item{chrSelect}{The chromosome used to train the partially methylated domains (PMDs) hidden Markov models. A shorter autosome is generally recommended. Default is chr22}

\item{umrs}{Logical. Whether or not to return UMRs. Default is FALSE.}

\item{nCores}{Number of cores to use for PMD-HMM training and LMR/UMR identification. Default is 1}
}
\description{
Identify low methylated regions (LMRs) (and unmethylated regions (UMRs)) for each sample from a bsseq object
}
\examples{
findLMRs(bs, methCutoff = 0.5, numCG = 4, chrSelect = "chr22", umrs = FALSE, nCores = 1)
}
