% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekers.R
\name{findSEMs}
\alias{findSEMs}
\title{Calculates CpG-level statistics and identifies stochastic epigenetic mutations (SEMs) from sequencing-based data}
\usage{
findSEMs(
  bs,
  type = c("raw", "smooth"),
  minCov = 10,
  minSamples = 50,
  runStats = FALSE,
  nCG = 1e+05,
  saveIndSEMs = FALSE,
  saveDir = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{bs}{A bsseq object}

\item{type}{Type of methylation values to be extracted from the bsseq object. Either raw or smooth values. For smoothed values, bsseq object needs to have been smoothed prior to running this option.}

\item{minCov}{Numeric value. The minimum number of reads to overlap a CpG sites by all samples for it to be used for downstream analysis. Note, lowering the number will likely increase the number of CpGs that are tested for SEM identification but at two big costs: 1) increase the computation time as it will be required to analyze more CpGs and 2) likely introduce false positives as less coverage generally equals less certainty of the estimation of methylation at a given CpG site. Default is 10.}

\item{minSamples}{Numeric value. The minimum number of samples required to perform the SEM analysis. A larger sample size (e.g., >50) is generally required for an analysis such as this. Default is 50.}

\item{runStats}{Logical. Whether to run CpG-level statistics (and save them) or not. Default is FALSE.}

\item{nCG}{Numeric. The number of CpGs used to split data into chunks for easier processing. Default is 100000 CpGs per chunk.}

\item{saveIndSEMs}{Logical. Whether sample-specific SEMs should be saved. Note, these files may be large in size, depending on how many SEMs are identified per sample. Default is FALSE.}

\item{saveDir}{Directory where outputs should be saved. Default is current working directory.}

\item{verbose}{Logical. Whether output of functions (e.g., number of CpGs assess per sample) should be verbose or not. Default is TRUE.}
}
\description{
Calculates CpG-level statistics and identifies stochastic epigenetic mutations (SEMs) from sequencing-based data
}
\examples{
findSEMs(bs, minCov = 10, minSamples = 50, runStats = FALSE, cluster = FALSE, nCG = 100000, saveIndSEMs = FALSE, saveDir = getwd(), verbose = TRUE)
}
