% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Return the current operating system}
\usage{
get_os()
}
\description{
A function that was shamelessly stolen from an r-bloggers posting \link{https://www.r-bloggers.com/identifying-the-os-from-r/}. It may someday be used to test for the operating system of a database user and utilize different \pkg{keyringr} functions based on the returned value.
}
\examples{
library(dplyr)
library(RPostgreSQL)
library(feather)

get_os()
}
\seealso{
\code{\link{db_connect}}
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
