% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvis_donut.R
\name{ggvis_donut}
\alias{ggvis_donut}
\title{Display a donut plot based on a proportion.}
\usage{
ggvis_donut(proportion, inner_radius = 50, outer_radius = 100,
  stroke = NA, fill.hover = c("#e24217", "#c8c6cc"), fill = c("#FC4A1A",
  "#dfdce3"))
}
\arguments{
\item{inner_radius}{The inner radius of the donut, in pixels. Defaults to 50.}

\item{outer_radius}{The outer radius of the donut, in pixels. Defaults to 100.}

\item{stroke}{The stroke color (i.e. outline) of the donut. Defaults to NA.}

\item{fill.hover}{The colors (from 'right' to 'left') of each segment of the donut to display when the mouse cursor hovers over a segment. Should be less than or equal to 2. Defaults to c("#e24217", '#c8c6cc').}

\item{fill}{The colors (from 'right' to 'left') of each segment of the donut to display when no mouse cursor is present. Should be less than or equal to 2. Defaults to c('#FC4A1A', '#dfdce3').}

\item{stroke}{The stroke color (i.e. outline) of the donut. Defaults to NA.}
}
\description{
Given a proportion (between 0 and 1), displays a donut plot using functions from the \pkg{ggvis} library.
}
\examples{
ggvis_donut(.5)
}
\seealso{
\code{\link{prop_to_radians}}
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
