% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{guessSex}
\alias{guessSex}
\title{Guesses the sex of samples from a bsseq object using the ratio of coverage from sex chromosomes, followed by k-means clustering}
\usage{
guessSex(bs)
}
\arguments{
\item{bs}{A bsseq object, must include CpGs on sex (X and Y) chromosomes.}
}
\description{
Guesses the sex of samples from a bsseq object using the ratio of coverage from sex chromosomes, followed by k-means clustering
}
\examples{
guessSex(bs)
}
