% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{guessSmoking}
\alias{guessSmoking}
\title{Guess smoking status (pack years), using R package wateRmelon to do the heavy lifting}
\usage{
guessSmoking(bs, method = NULL)
}
\arguments{
\item{bs}{A bsseq object. Removing low-coverage CpGs is not required}

\item{method}{Method used for calculation. Please reference wateRmelon::smokp() methods for the available options. Default is Packyears}
}
\description{
Guess smoking status (pack years), using R package wateRmelon to do the heavy lifting
}
\examples{
guessSmoking(bs)
}
