% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{imprintMe}
\alias{imprintMe}
\title{Overlaps DMRs with known imprinted regions}
\usage{
imprintMe(dmrs)
}
\arguments{
\item{dmrs}{A GenomicRanges object containing the chromosome, start, and end of regions of interest (e.g., differentially methylated regions (DMRs)), with coordinates corresponding to the hg38 version of the human genome}
}
\description{
Overlaps DMRs with known imprinted regions
}
\examples{
imprintMe(dmrs)
}
