% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inverse.transform}
\alias{inverse.transform}
\title{Inverse transformation of predicted age}
\usage{
inverse.transform(tAge, adult.age = 20)
}
\arguments{
\item{tAge}{Transformed age from prediction model}

\item{adult.age}{Adult age for transformation. Default is 20}
}
\description{
Inverse transformation of predicted age
}
\examples{
inverse.transform(bs, adult.age = 20)
}
