% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{liftBuild}
\alias{liftBuild}
\title{Convert coordinates from hg19 (or hg38) to hg38 (or hg19)}
\usage{
liftBuild(bs, current = c("hg19", "hg38"), new = c("hg19", "hg38"))
}
\arguments{
\item{bs}{A bsseq object}

\item{current}{The current human genome build. Can be one of hg19 or hg38}

\item{new}{The human genome build that you would like to convert to. Can be one of hg19 or hg38}
}
\description{
Convert coordinates from hg19 (or hg38) to hg38 (or hg19)
}
\examples{
liftBuild(bs, current = "hg38", new = "hg19")
}
