% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_flat_files.R
\name{load_flat_files}
\alias{load_flat_files}
\title{Load data from \code{oliver_replica} into workspace}
\usage{
load_flat_files(tables = c("report_businessmetrics_memberbreakdown",
  "report_businessmetrics_provideractivity", "report_businessmetrics_rollup",
  "report_pcvmetrics_servicereferrals", "report_pcvmetrics_visitreports",
  "Organizations"))
}
\arguments{
\item{tables}{A list of feather tables that which have been written to the \code{Data} directory with \code{\link{create_flat_files}}.
defaults to c('report_businessmetrics_memberbreakdown', 'report_businessmetrics_provideractivity', 'report_businessmetrics_rollup'
,'report_pcvmetrics_servicereferrals', 'report_pcvmetrics_visitreports', 'Organizations')}
}
\value{
\eqn{n} tibbles/data frames where \eqn{n} is equal to \code{length(tables)}.
}
\description{
Takes a vector of feather tables created with \code{\link{create_flat_files}} and loads them into \code{.GlobalEnv}.
}
\examples{
load_flat_files()
}
\seealso{
\code{\link{create_flat_files}}
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
