% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{madrid}
\alias{madrid}
\title{Predicts the age of samples from a bsseq object}
\usage{
madrid(
  bs,
  adult.age = 20,
  checkCov = TRUE,
  fastImpute = FALSE,
  imp = 5,
  calcSeqClock = TRUE,
  calcArrayClocks = TRUE,
  gestationalAge = FALSE,
  offset = 0
)
}
\arguments{
\item{bs}{A bsseq object}

\item{adult.age}{Adult age for transformation. Default is 20}

\item{checkCov}{Logical. Whether to check coverage of samples to ensure sufficient coverage is present to estimate age. Default is TRUE}

\item{fastImpute}{Logical. Whether fast imputation should be performed. This is not advised as it leads to less accurate predictions. Default is FALSE}

\item{imp}{Number of imputations to be performed by mice() during imputation of missing methylation values. Default is 5}

\item{calcSeqClock}{Logical. Whether sequencing-based clock (MADRID) should be calculated. Default is TRUE.}

\item{calcArrayClocks}{Logical. Whether array-based clocks (Horvath, Hannum, PhenoAge, SkinBlood, Lin) should be calculated. Default is TRUE.}

\item{gestationalAge}{Logical. Whether array-based clocks of gestational age should be calculated. Default is FALSE}

\item{offset}{Integer. The number of basepairs to add or subtract from your CpG coordinates. Depends on whether the genome assembly of your data was zero-based or one-based. Coordinates used in this function are one-based. Default is 0.}
}
\description{
Predicts the age of samples from a bsseq object
}
\examples{
madrid(bs, adult.age = 20, fastImpute = FALSE, imp = 5, calcSeqClock = TRUE, calcArrayClocks = TRUE, gestationalAge = FALSE, offset = 0)
}
