% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_value_raw.R
\name{measurement_value_raw}
\alias{measurement_value_raw}
\title{Capture the "raw value"}
\usage{
measurement_value_raw(id, measurement_service_val = NA,
  measurement_index_val = NA)
}
\arguments{
\item{id}{The provider \code{id} as defined in \code{oliver_replica}.}
}
\description{
This function is designed to be used as a method within an object of class \code{performance_metric_provider}.
It assumes the existence of #' a file named \code{network_aggr_performance_metrics.feather} as created by \code{\link{write_network_metric_file}}. It returns the raw value of a given performance measure, as determined by the column name in \code{network_aggr_performance_metrics.feather}.
}
