% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oliver_replica_connect.R
\name{oliver_replica_connect}
\alias{oliver_replica_connect}
\title{Open a connection to the \code{oliver_replica} database.}
\usage{
oliver_replica_connect(database_name = "oliver_replica",
  database_host = "127.0.0.1", database_user = "mienkoja")
}
\arguments{
\item{database_name}{A character string specifying the name of the database. Defaults to \code{oliver_replica}.}

\item{database_host}{A character string specifying the name of the database host. Defaults to \code{localhost}.}

\item{database_user}{A character string specifying the name of the user. Defaults to \code{mienkoja}.}
}
\value{
A list of class \code{PostgreSQLConnection}.
}
\description{
This function is used to open a connection to the \code{oliver_replica} database and set the database search path to \code{staging} as suggested here \link{https://github.com/tidyverse/dplyr/issues/244}. This function has only been tested on macOS. It is probably pretty fragile.
}
\details{
#' The following dependencies are required in order to create a valid connection:

#' \itemize{
  \item The presence of the \code{oliver_replica} database on \code{localhost}, possibly via SSH as partially decribed in the \code{\link{create_ssh_command}} documentation.
  \item A valid username and password on the \code{oliver_replica} database. Can be set by contacting \email{mattbro@uw.edu}.
  \item A password stored on your Windows, Linux, or mac box for use in the \pkg{keyringr} package. Parameters set on your local machine should follow the format \code{oliver_replica_myuser} where \code{myuser} is replaced with the userid. 
}
}
\examples{

db_connect()
}
\seealso{
\code{\link{create_ssh_command}}
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
