% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_target_raw.R
\name{percent_target_raw}
\alias{percent_target_raw}
\title{Capture the "percent conforming"}
\usage{
percent_target_raw(id, measurement_service_val = NA,
  measurement_index_val = NA)
}
\arguments{
\item{id}{The provider \code{id} as defined in \code{oliver_replica}.}
}
\description{
This function is designed to be used as a method within an object of class \code{performance_metric_provider}.
It assumes the existence of #' a file named \code{network_aggr_performance_metrics.feather} as created by \code{\link{write_network_metric_file}}. It returns the "percent conforming" to a specified target of a given performance measure, as determined by the column name in \code{network_aggr_performance_metrics.feather}.
}
