% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_to_radians.R
\name{prop_to_radians}
\alias{prop_to_radians}
\title{Convert a proportion to radians}
\usage{
prop_to_radians(proportion)
}
\description{
A function that converts a proportion to radians as required by some of the parameters of \code{layer_arcs} in \pkg{ggvis}.
}
\examples{

prop_to_radians(.5) == pi

}
\seealso{
\code{\link{ggvis_donut}}
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
