% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Parse SQL from file.}
\usage{
read_sql(filename, silent = TRUE)
}
\arguments{
\item{filename}{A connection object or a character string.}

\item{silent}{logical. Warn if a text file is missing a final EOL or if there are embedded nulls. Defaults to TRUE.}
}
\value{
A character vector of length the number of lines read.
}
\description{
Parse raw SQL from a file. The main purpose of this function is to get rid of comments which R does not 
handle very well within the DBI or RODBC packages (or their wrappers (e.g. dplyr)).
}
\examples{
fname <- tempfile("sql",fileext=".sql")
cat("--query 6
    select a6.column1, --trailing comments
    a6.column2, ---test triple -
    count(a6.column3) as counts, --/* funny comment */
    a6.column3 - a6.column4 ---test single - 
    /*count the number of occurences in table 1; 
    test another comment style
    */
    from data.table a6 /* --1st weirdo comment */
    /* --2nd weirdo comment */
    group by a6.column1\\n", file=fname)
read_sql(fname)
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
