% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_for_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data for analysis}
\format{dataset with 152 observation and 14 columns
\describe{
\item{Sample_ID}{column with the name of every sample}
\item{Label}{Label of the geographic site of sampling}
}}
\usage{
sample_data
}
\description{
load data
}
\keyword{datasets}
