% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{transform_age}
\alias{transform_age}
\title{Transformation of actual age}
\usage{
transform_age(age, adult.age = 20)
}
\arguments{
\item{age}{Age of sample(s)}

\item{adult.age}{Adult age for transformation. Default is 20}
}
\description{
Transformation of actual age
}
\examples{
transform_age(age, adult.age = 20)
}
