% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_network_metric_file.R
\name{write_network_metric_file}
\alias{write_network_metric_file}
\title{Combine data from \code{oliver_replica} into a network metric file}
\usage{
write_network_metric_file(time_window_days = 30,
  network_providers = c("American Indian Community Center",
  "Angela's Family Services", "Fulcrum Institute, DRC", "G&A, LLC",
  "GMC Training Institute", "Love My Child Family Services",
  "Martin Luther King Jr. Family Outreach Center",
  "New Beginnings Visitation and Parent Education", "Reunified Services",
  "Service Alternatives",
  "The Salvation Army Nurturing Center for Children & Families",
  "Unite Family Services", "Family Visitation Volunteers of America"),
  set_neg_to_zero = FALSE, set_na_to_avg = FALSE)
}
\arguments{
\item{time_window_days}{numeric window of time over which measures are to be averaged.}

\item{network_providers}{vector of providers to be included in the report. Defaults to current Family Impact Network providers as shown above.}

\item{set_neg_to_zero}{logical scalar - Should the measurements that are less than 0 be set to 0? Defaults to \code{FALSE}. Experimental.}

\item{set_na_to_avg}{logical scalar - Should the measurements that are \code{NA} or \code{NaN} be set to the network mean? Defaults to \code{FALSE}. Experimental.}
}
\value{
A single feather file.
}
\description{
Takes tables loaded with \code{\link{load_flat_files}}, aggregates them to provider-level metrics within a given network, and
writes the file to the \code{Data} directory of the package.
}
\examples{
write_network_metric_file()
}
\seealso{
\code{\link{load_flat_files}}
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
