% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_oliver_data.R
\name{write_oliver_data}
\alias{write_oliver_data}
\title{Get data from \code{oliver_replica} write a single file}
\usage{
write_oliver_data(data_name, con, schema)
}
\arguments{
\item{data_name}{A character string specifying the name of the table or view being accessed from the oliver replica.}

\item{con}{An object of class \code{src_postgres} which specifies a connection to the oliver replica.}

\item{schema}{A character string specifying the name of the schema holding the table or view from the oliver replica. The default value is \code{static}.}
}
\value{
A single feather file.
}
\description{
This function is used to select all records from a view or a table within the specified schema within \code{oliver_replica}. For now, the files are written in a \pkg{feather} format in the \code{Data} directory of the package.
}
\details{
#' The following dependencies are required in order to create a valid string:

#' \itemize{
  \item The presence of an open SSH tunnel to the oliver replica as partially decribed in the \code{\link{create_ssh_command}} documentation.
  \item A valid username and password on the oliver replica database. Can be set by contacting \email{mattbro@uw.edu}.
}
}
\examples{
library(dplyr)
library(RPostgreSQL)
library(feather)

con <- src_postgres(
  dbname = 'oliver_replica',
  host = '127.0.0.1',
  user = 'mienkoja',
  password = 'mypassword'
)

write_oliver_data(data_name = 'report_businessmetrics_memberbreakdown', con = con)
}
\seealso{
\code{\link{create_ssh_command}}, \code{\link{create_flat_files}}
}
\author{
Joe Mienko, \email{mienkoja@uw.edu}
}
