% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{madrid}
\alias{madrid}
\title{Predicts the age of samples from a bsseq object}
\usage{
madrid(
  bs,
  adult.age = 20,
  fastImpute = FALSE,
  imp = 5,
  calcSeqClock = TRUE,
  calcArrayClocks = TRUE
)
}
\arguments{
\item{bs}{A bsseq object}

\item{adult.age}{Adult age for transformation. Default is 20}

\item{fastImpute}{Logical. Whether fast imputation should be performed. This is not advised as it leads to less accurate predictions. Default is FALSE}

\item{imp}{Number of imputations to be performed by mice() during imputation of missing methylation values. Default is 5}

\item{calcSeqClock}{Logical. Whether sequencing-based clock (MADRID) should be calculated. Default is TRUE.}

\item{calcArrayClocks}{Logical. Whether array-based clocks (Horvath, Hannum, PhenoAge, SkinBlood, Lin) should be calculated. Default is TRUE.}
}
\description{
Predicts the age of samples from a bsseq object
}
\examples{
madrid(bs, adult.age = 20, fastImpute = FALSE, imp = 5, calcSeqClock = TRUE, calcArrayClocks = TRUE)
}
