% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOliveR.R
\name{runOliveR}
\alias{runOliveR}
\title{Launches the web Shiny app "OliveR"}
\usage{
runOliveR()
}
\arguments{
\item{host}{optional parameter where to execute the application, default is 127.0.0.1}

\item{ui}{R script containing the code of the user interface of the shiny app}

\item{server}{R script containing the code of the server of the shiny app}
}
\value{
The function loads the web application OliveR in your default browser
through the \code{\link{shinyApp}} function, which requires a script with the
User Interface of the app and a script of the Server.
}
\description{
Best package in the world
}
\examples{
\dontrun{
runOliver()
}
}
